Refactor competition reading and processing to its own file
Move competition reading JSON and processing code to a new file `competition.js` and clean up the code.

* **New File `competition.js`**
  - Create a new file to handle competition-related logic.
  - Move the code for reading and processing the competition JSON from `server.js` to this file.
  - Move the code for handling file uploads and parsing Lenex files from `server.js` to this file.
  - Move the code for handling competition-related endpoints from `server.js` to this file.
  - Move the code for finding athletes in the competition data from `server.js` to this file.

* **Modified `server.js`**
  - Remove the code for reading and processing the competition JSON.
  - Remove the code for handling file uploads and parsing Lenex files.
  - Remove the code for handling competition-related endpoints.
  - Remove the code for finding athletes in the competition data.
  - Import and use the functions from `competition.js`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/spelbreker/ws-swim-stopwatch/tree/add-lenex-upload?shareId=XXXX-XXXX-XXXX-XXXX).
